test: Move integration tests from unit tests module to appropriate directory
