+++ moved/tests/config.sh
