+++ moved/tests/res/ERC20.json
