+++ moved/tests/res/counter/Move.toml
