+++ moved/tests/res/counter/sources/counter.move
