+++ moved/tests/res/signer-struct/sources/signer_struct.move
